+++ src/app/views/Logout.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-  import {tryFunc} from "hurdak"
   import {fly} from "src/util/transition"
   import Anchor from "src/partials/Anchor.svelte"
   import Content from "src/partials/Content.svelte"
@@ -10,7 +9,7 @@
   const confirm = async () => {
     confirmed = true
 
-    await tryFunc(() => storage.clear())
+    storage.clear()
 
     // do a hard refresh so everything gets totally cleared.
     // Give them a moment to see the state transition. IndexedDB